backend/requirements.txt
pandas==2.2.3
pdf2image==1.17.0
poppler-utils==0.1.0
Flask==3.1.0
flask-cors==5.0.1
relevanceai==10.2.2
azure-core==1.32.0
azure-ai-vision-imageanalysis==1.0.0
python-dotenv==1.0.1
openpyxl==3.1.5
pypdf==5.4.0
pytesseract==0.3.13
pdf2image==1.17.0
openai==1.69.0